# backend/tool-integration/openapi.yaml
openapi: 3.0.3
info:
  title: VidSage User Management & Google OAuth API
  description: |
    REST API for managing Google OAuth 2.0 authentication and Google Docs integration.

    ## Features
    - Google OAuth 2.0 Authorization Code flow
    - Automatic token refresh with offline access
    - Create and manage Google Docs with AI-generated content
    - Secure encrypted token storage in MongoDB
    - Clerk user ID integration

    ## Authentication Flow
    1. Initiate OAuth: `GET /auth/google?userId={clerkUserId}`
    2. User grants consent on Google
    3. Callback handles token exchange: `GET /auth/google/callback`
    4. Tokens stored securely, linked to Clerk user ID
    5. Create Google Docs: `POST /google/docs`

    ## Security
    - Tokens encrypted at rest using AES-256
    - OAuth state parameter for CSRF protection
    - Automatic token refresh before expiration
    - HTTPS required in production

  version: 1.0.0
  contact:
    name: VidSage API Support
    email: support@vidsage.com
  license:
    name: ISC
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.vidsage.com
    description: Production server
...
paths:
...
  /google/status:
    get:
      tags:
        - OAuth
      summary: Check user's Google OAuth status
      description: |
        Check if a user has authorized Google access and token status.

        **Use Cases:**
        - Before showing "Connect Google" vs "Save to Google Docs" button
        - Checking if re-authentication is needed
        - Verifying token validity
      operationId: getGoogleOAuthStatus
      parameters:
        - name: userId
          in: query
          required: true
          description: Clerk user ID
          schema:
            type: string
            example: user_2abcd1234xyz
      responses:
        "200":
          description: OAuth status information
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorized:
                    type: boolean
                    description: Whether user has authorized Google access
                  hasRefreshToken:
                    type: boolean
                    description: Whether refresh token is available
                  tokenExpired:
                    type: boolean
                    nullable: true
                    description: Whether access token is expired (null if not authorized)
              example:
                authorized: true
                hasRefreshToken: true
                tokenExpired: false
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Human-readable error message
          example: userId and content are required
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: object
          description: Additional error details (development only)
          nullable: true
...
  securitySchemes:
    ClerkAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Clerk session token for user authentication.

        **Format:** `Bearer <clerk_session_token>`

        **Note:** Currently, userId is passed in request body/query.
        In production, implement Clerk token verification middleware.
security:
  - ClerkAuth: []
...